LPD-54027 Rename services folder
