+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/actions/deleteStructureAction.ts
@@ -6,7 +6,7 @@
 import {openModal} from 'frontend-js-components-web';
 import {sub} from 'frontend-js-web';
 
-import {fetchJSON} from '../../../api/api';
+import {fetchJSON} from '../../../services/api';
 import DeleteStructureModalContent from '../../components/modal/DeleteStructureModalContent';
 import {executeAsyncItemAction} from '../utils/executeAsyncItemAction';
