+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/services/CategoryService.ts
@@ -5,7 +5,7 @@
 
 import {fetch} from 'frontend-js-web';
 
-import {HEADERS_ALL_LANGUAGES} from '../../../api/api';
+import {HEADERS_ALL_LANGUAGES} from '../../../services/api';
 
 const createCategory = async (
 	categoryByVocabularyIdApiUrl: string,
